More changes to hooks.
Separated Hooks and Signal so they are not derived from the other.

Provided a process method which will yield the result of each called
callback.  This can be used in place of get to avoid having a reference
to a callback when the parameters are the same for each call:

    Instead of:

        sum = 0
        for cb in h.get("name"):
            sum += cb(10)

    You can do:

        sum = 0
        for value in h.process("name", 10):
            sume += value

The get method is still available when it is needed, such as calling
callbacks with different values.  It is also possible to terminal the
callbacks early from get can process simply by not completing the
iterations.